requirements.windows.txt
pywin32
ghostscript